Simplify organization of base classes (#675)
* Make _check_ncores a function instead of a method.

* Drop Transformer base class.

All it had was an abstract method that was always overridden.

* Move Decoder to new decode/base.py.

* Create IBMAEstimator base class.

* Move CBMA stuff from MetaEstimator to CBMAEstimator.

* Remove MetaEstimator.

* Make kwargs more explicit.

* Replace _validate_input with _collect_inputs.

* Fix import.

* Fix things.

* Update ibma.py

* Update api.rst

* Add version directives.

* Update base.py

* Make some methods abstract.

Specifically, CBMAEstimator._compute_summarystat_est and Estimator._preprocess_input.

* Fix ALESubtraction.

* Fix MKDAChi2.
