+++ nimare/diagnostics.py
@@ -13,7 +13,7 @@ from scipy.spatial.distance import cdist
 from tqdm.auto import tqdm
 
 from nimare.base import NiMAREBase
-from nimare.utils import mm2vox, tqdm_joblib, vox2mm
+from nimare.utils import _check_ncores, mm2vox, tqdm_joblib, vox2mm
 
 LGR = logging.getLogger(__name__)
 
@@ -60,7 +60,7 @@ class Jackknife(NiMAREBase):
     ):
         self.target_image = target_image
         self.voxel_thresh = voxel_thresh
-        self.n_cores = self._check_ncores(n_cores)
+        self.n_cores = _check_ncores(n_cores)
 
     def transform(self, result):
         """Apply the analysis to a MetaResult.
@@ -278,7 +278,7 @@ class FocusCounter(NiMAREBase):
     ):
         self.target_image = target_image
         self.voxel_thresh = voxel_thresh
-        self.n_cores = self._check_ncores(n_cores)
+        self.n_cores = _check_ncores(n_cores)
 
     def transform(self, result):
         """Apply the analysis to a MetaResult.
