+++ nimare/meta/cbma/ale.py
@@ -12,7 +12,7 @@ from nimare.meta.cbma.base import CBMAEstimator, PairwiseCBMAEstimator
 from nimare.meta.kernel import ALEKernel
 from nimare.stats import null_to_p, nullhist_to_p
 from nimare.transforms import p_to_z
-from nimare.utils import tqdm_joblib, use_memmap
+from nimare.utils import _check_ncores, tqdm_joblib, use_memmap
 
 LGR = logging.getLogger(__name__)
 
@@ -146,7 +146,7 @@ class ALE(CBMAEstimator):
         super().__init__(kernel_transformer=kernel_transformer, **kwargs)
         self.null_method = null_method
         self.n_iters = n_iters
-        self.n_cores = self._check_ncores(n_cores)
+        self.n_cores = _check_ncores(n_cores)
         self.dataset = None
 
     def _compute_summarystat_est(self, ma_values):
@@ -332,7 +332,7 @@ class ALESubtraction(PairwiseCBMAEstimator):
         self.dataset1 = None
         self.dataset2 = None
         self.n_iters = n_iters
-        self.n_cores = self._check_ncores(n_cores)
+        self.n_cores = _check_ncores(n_cores)
         # memory_limit needs to exist to trigger use_memmap decorator, but it will also be used if
         # a Dataset with pre-generated MA maps is provided.
         self.memory_limit = "100mb"
@@ -357,8 +357,8 @@ class ALESubtraction(PairwiseCBMAEstimator):
         n_grp1, n_voxels = ma_maps1.shape
 
         # Get ALE values for the two groups and difference scores
-        grp1_ale_values = 1.0 - np.prod(1.0 - ma_maps1, axis=0)
-        grp2_ale_values = 1.0 - np.prod(1.0 - ma_maps2, axis=0)
+        grp1_ale_values = self._compute_summarystat_est(ma_maps1)
+        grp2_ale_values = self._compute_summarystat_est(ma_maps2)
         diff_ale_values = grp1_ale_values - grp2_ale_values
         del grp1_ale_values, grp2_ale_values
 
@@ -425,6 +425,10 @@ class ALESubtraction(PairwiseCBMAEstimator):
             "logp_desc-group1MinusGroup2": logp_arr,
         }
         return images
+
+    def _compute_summarystat_est(self, ma_values):
+        stat_values = 1.0 - np.prod(1.0 - ma_values, axis=0)
+        return stat_values
 
     def _run_permutation(self, i_iter, n_grp1, ma_arr, iter_diff_values):
         """Run a single permutations of the ALESubtraction null distribution procedure.
@@ -566,7 +570,7 @@ class SCALE(CBMAEstimator):
 
         self.xyz = xyz
         self.n_iters = n_iters
-        self.n_cores = self._check_ncores(n_cores)
+        self.n_cores = _check_ncores(n_cores)
         # memory_limit needs to exist to trigger use_memmap decorator, but it will also be used if
         # a Dataset with pre-generated MA maps is provided.
         self.memory_limit = "100mb"
